Changed How Ability & Activations Work
Changed and edited scripts so does not matter what power you select in the action bar it will activate that the predefined data is on the power container.  Also made it so when you learn a power it adds it to the action bar and caps it to only the 5 slots on the inventory.  Cooldown for the power has also been added so u can't mash it and cause the game to slow down.