Added Inventory + Modular Quest System
Added a modular quest system that gives a full list of customizable ways to make quests and dialog and rewards in the game. Added a inventory system that dynamically adds in items you pickup into the inventory and has a scriptable data for every item so can know how many it is and stackable or not and quest item or not etc.

Added shoes as a starter quest item.

Bugs with the dynamic inventory system due to removing item from game world too quickly.